feat: ensure ServerResponse is always patched
When http is disabled, the writeHead patch on ServerResponse was
not applied on https servers. With this change, both http and https
will attempt to patch it independently and it will keep track of
if it has been patched so it doesn't duplicate the patch.
