+++ lib/instrumentation/http-shared.js
@@ -4,6 +4,7 @@ var url = require('url')
 
 var endOfStream = require('end-of-stream')
 
+var shimmer = require('./shimmer')
 var symbols = require('../symbols')
 
 exports.instrumentRequest = function (agent, moduleName) {
@@ -170,4 +171,44 @@ exports.traceOutgoingRequest = function (agent, moduleName) {
       }
     }
   }
+}
+
+var writeHeadIsWrapped = false
+exports.instrumentWriteHead = function (agent, ServerResponse) {
+  if (writeHeadIsWrapped) return
+  if (!ServerResponse) return
+
+  writeHeadIsWrapped = true
+  agent.logger.debug('shimming ServerResponse.prototype.writeHead function')
+  shimmer.wrap(ServerResponse.prototype, 'writeHead', original => {
+    return function wrappedWriteHead () {
+      var headers = arguments.length === 1
+        ? this._headers // might be because of implicit headers
+        : arguments[arguments.length - 1]
+
+      var result = original.apply(this, arguments)
+
+      var trans = this[symbols.requestTransaction]
+      this[symbols.requestTransaction] = null
+
+      if (trans) {
+        // It shouldn't be possible for the statusCode to be falsy, but just in
+        // case we're in a bad state we should avoid throwing
+        trans.result = 'HTTP ' + (this.statusCode || '').toString()[0] + 'xx'
+
+        // End transacton early in case of SSE
+        if (headers && typeof headers === 'object' && !Array.isArray(headers)) {
+          Object.keys(headers).some(function (key) {
+            if (key.toLowerCase() !== 'content-type') return false
+            if (String(headers[key]).toLowerCase().indexOf('text/event-stream') !== 0) return false
+            agent.logger.debug('detected SSE response - ending transaction %o', { id: trans.id })
+            agent.endTransaction()
+            return true
+          })
+        }
+      }
+
+      return result
+    }
+  })
 }
