+++ lib/instrumentation/modules/http.js
@@ -2,7 +2,6 @@
 
 var httpShared = require('../http-shared')
 var shimmer = require('../shimmer')
-var symbols = require('../../symbols')
 
 module.exports = function (http, agent, version, enabled) {
   var disableTransactions = agent._conf.disableTransactions
@@ -10,8 +9,7 @@ module.exports = function (http, agent, version, enabled) {
     agent.logger.debug('shimming http.Server.prototype.emit function')
     shimmer.wrap(http && http.Server && http.Server.prototype, 'emit', httpShared.instrumentRequest(agent, 'http'))
 
-    agent.logger.debug('shimming http.ServerResponse.prototype.writeHead function')
-    shimmer.wrap(http && http.ServerResponse && http.ServerResponse.prototype, 'writeHead', wrapWriteHead)
+    httpShared.instrumentWriteHead(agent, http && http.ServerResponse)
   }
 
   if (!enabled) return http
@@ -20,36 +18,4 @@ module.exports = function (http, agent, version, enabled) {
   shimmer.wrap(http, 'request', httpShared.traceOutgoingRequest(agent, 'http'))
 
   return http
-
-  function wrapWriteHead (original) {
-    return function wrappedWriteHead () {
-      var headers = arguments.length === 1
-        ? this._headers // might be because of implicit headers
-        : arguments[arguments.length - 1]
-
-      var result = original.apply(this, arguments)
-
-      var trans = this[symbols.requestTransaction]
-      this[symbols.requestTransaction] = null
-
-      if (trans) {
-        // It shouldn't be possible for the statusCode to be falsy, but just in
-        // case we're in a bad state we should avoid throwing
-        trans.result = 'HTTP ' + (this.statusCode || '').toString()[0] + 'xx'
-
-        // End transacton early in case of SSE
-        if (headers && typeof headers === 'object' && !Array.isArray(headers)) {
-          Object.keys(headers).some(function (key) {
-            if (key.toLowerCase() !== 'content-type') return false
-            if (String(headers[key]).toLowerCase().indexOf('text/event-stream') !== 0) return false
-            agent.logger.debug('detected SSE response - ending transaction %o', { id: trans.id })
-            agent.endTransaction()
-            return true
-          })
-        }
-      }
-
-      return result
-    }
-  }
 }
