+++ lib/instrumentation/modules/https.js
@@ -10,6 +10,8 @@ module.exports = function (https, agent, version, enabled) {
   if (!disableTransactions.has('https')) {
     agent.logger.debug('shimming https.Server.prototype.emit function')
     shimmer.wrap(https && https.Server && https.Server.prototype, 'emit', httpShared.instrumentRequest(agent, 'https'))
+
+    httpShared.instrumentWriteHead(agent, http && http.ServerResponse)
   }
 
   if (!enabled) return https
